% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monti-package.R
\docType{package}
\name{monti-package}
\alias{monti}
\alias{monti-package}
\title{monti: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\keyword{internal}
