% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSML-package.R
\docType{package}
\name{SSML-package}
\alias{SSML}
\alias{SSML-package}
\title{SSML: Semi-Supervised Machine Learning Methods for Bioconductor}
\description{
SSML aims to impliment semi-supervised machine learning methods for
    Bioconductor SummarizedExperiment class objects to predict labels of
    carcinogenicity. However, we will write our methods to be sufficiently
    general so as to apply to additional problems.
}
\keyword{internal}
