% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSML-package.R
\docType{package}
\name{SSML-package}
\alias{SSML}
\alias{SSML-package}
\title{SSML: What the Package Does (One Line, Title Case)}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis tempor massa
    eu ipsum interdum, ut tincidunt mauris efficitur. Maecenas placerat pulvinar
    egestas. Maecenas facilisis interdum aliquam. Phasellus sit amet pharetra
    eros, vel ullamcorper sapien. Curabitur cursus orci justo, ultrices rhoncus
    magna tincidunt eu. Mauris nec gravida orci. Nulla ac enim quam. Maecenas
    eros risus, varius nec nisi ac, varius ultricies massa.
}
\keyword{internal}
